backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
groq==0.4.1
python-dotenv==1.0.0
Werkzeug==3.0.1